Don't checkout any revision when cloning.
This guards us against git refusing to checkout the given tag due to
oddities in the repo. For instance, Drupal 8-beta11ish managed to commit
a binary file with line ending conversion, which causes the file to be
locally modified on checkout. That would have broken core support until
they fixed it, without this change.